backend/pyproject.toml
[project]
name = "multimodal-rag-backend"
version = "0.1.0"
description = "Backend for multimodal RAG system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "llama-index>=0.12.0",
    "llama-index-embeddings-ollama>=0.5.0",
    "llama-index-llms-ollama>=0.5.0",
    "qdrant-client>=1.12.0",
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "watchdog>=6.0.0",
    "pymupdf>=1.25.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "python-pptx>=1.0.0",
    "easyocr>=1.7.0",
    "faster-whisper>=1.1.0",
    "markitdown>=0.0.1a3",
    "av>=13.0.0",
    "watchdog>=6.0.0",
    "sentence-transformers>=3.3.0",
    "transformers>=4.46.0",
    "torch>=2.5.0",
    "numpy>=1.26.0",
    "pillow>=11.0.0",
    "aiofiles>=24.1.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
]
